+++ src/Executor/Values.php
@@ -67,48 +67,9 @@ class Values
             /** @var InputType|Type $varType */
             $varType = TypeInfo::typeFromAST($schema, $varDefNode->type);
 
-            if (Type::isInputType($varType)) {
-                if (array_key_exists($varName, $inputs)) {
-                    $value   = $inputs[$varName];
-                    $coerced = Value::coerceValue($value, $varType, $varDefNode);
-                    /** @var Error[] $coercionErrors */
-                    $coercionErrors = $coerced['errors'];
-                    if (empty($coercionErrors)) {
-                        $coercedValues[$varName] = $coerced['value'];
-                    } else {
-                        $messagePrelude = sprintf(
-                            'Variable "$%s" got invalid value %s; ',
-                            $varName,
-                            Utils::printSafeJson($value)
-                        );
-
-                        foreach ($coercionErrors as $error) {
-                            $errors[] = new Error(
-                                $messagePrelude . $error->getMessage(),
-                                $error->getNodes(),
-                                $error->getSource(),
-                                $error->getPositions(),
-                                $error->getPath(),
-                                $error,
-                                $error->getExtensions()
-                            );
-                        }
-                    }
-                } else {
-                    if ($varType instanceof NonNull) {
-                        $errors[] = new Error(
-                            sprintf(
-                                'Variable "$%s" of required type "%s" was not provided.',
-                                $varName,
-                                $varType
-                            ),
-                            [$varDefNode]
-                        );
-                    } elseif ($varDefNode->defaultValue !== null) {
-                        $coercedValues[$varName] = AST::valueFromAST($varDefNode->defaultValue, $varType);
-                    }
-                }
-            } else {
+            if (! Type::isInputType($varType)) {
+                // Must use input types for variables. This should be caught during
+                // validation, however is checked again here for safety.
                 $errors[] = new Error(
                     sprintf(
                         'Variable "$%s" expected value of type "%s" which cannot be used as an input type.',
@@ -117,6 +78,61 @@ class Values
                     ),
                     [$varDefNode->type]
                 );
+            } else {
+                $hasValue = array_key_exists($varName, $inputs);
+                $value    = $hasValue ? $inputs[$varName] : Utils::undefined();
+
+                if (! $hasValue && $varDefNode->defaultValue) {
+                    // If no value was provided to a variable with a default value,
+                    // use the default value.
+                    $coercedValues[$varName] = AST::valueFromAST($varDefNode->defaultValue, $varType);
+                } elseif ((! $hasValue || $value === null) && ($varType instanceof NonNull)) {
+                    // If no value or a nullish value was provided to a variable with a
+                    // non-null type (required), produce an error.
+                    $errors[] = new Error(
+                        sprintf(
+                            $hasValue
+                                ? 'Variable "$%s" of non-null type "%s" must not be null.'
+                                : 'Variable "$%s" of required type "%s" was not provided.',
+                            $varName,
+                            Utils::printSafe($varType)
+                        ),
+                        [$varDefNode]
+                    );
+                } elseif ($hasValue) {
+                    if ($value === null) {
+                        // If the explicit value `null` was provided, an entry in the coerced
+                        // values must exist as the value `null`.
+                        $coercedValues[$varName] = null;
+                    } else {
+                        // Otherwise, a non-null value was provided, coerce it to the expected
+                        // type or report an error if coercion fails.
+                        $coerced = Value::coerceValue($value, $varType, $varDefNode);
+                        /** @var Error[] $coercionErrors */
+                        $coercionErrors = $coerced['errors'];
+                        if ($coercionErrors) {
+                            $messagePrelude = sprintf(
+                                'Variable "$%s" got invalid value %s; ',
+                                $varName,
+                                Utils::printSafeJson($value)
+                            );
+
+                            foreach ($coercionErrors as $error) {
+                                $errors[] = new Error(
+                                    $messagePrelude . $error->getMessage(),
+                                    $error->getNodes(),
+                                    $error->getSource(),
+                                    $error->getPositions(),
+                                    $error->getPath(),
+                                    $error,
+                                    $error->getExtensions()
+                                );
+                            }
+                        } else {
+                            $coercedValues[$varName] = $coerced['value'];
+                        }
+                    }
+                }
             }
         }
 
@@ -208,27 +224,32 @@ class Values
             $argType           = $argumentDefinition->getType();
             $argumentValueNode = $argumentValueMap[$name] ?? null;
 
-            if ($argumentValueNode === null) {
-                if ($argumentDefinition->defaultValueExists()) {
-                    $coercedValues[$name] = $argumentDefinition->defaultValue;
-                } elseif ($argType instanceof NonNull) {
+            if ($argumentValueNode instanceof VariableNode) {
+                $variableName = $argumentValueNode->name->value;
+                $hasValue     = $variableValues ? array_key_exists($variableName, $variableValues) : false;
+                $isNull       = $hasValue ? $variableValues[$variableName] === null : false;
+            } else {
+                $hasValue = $argumentValueNode !== null;
+                $isNull   = $argumentValueNode instanceof NullValueNode;
+            }
+
+            if (! $hasValue && $argumentDefinition->defaultValueExists()) {
+                // If no argument was provided where the definition has a default value,
+                // use the default value.
+                $coercedValues[$name] = $argumentDefinition->defaultValue;
+            } elseif ((! $hasValue || $isNull) && ($argType instanceof NonNull)) {
+                // If no argument or a null value was provided to an argument with a
+                // non-null type (required), produce a field error.
+                if ($isNull) {
                     throw new Error(
-                        'Argument "' . $name . '" of required type ' .
-                        '"' . Utils::printSafe($argType) . '" was not provided.',
+                        'Argument "' . $name . '" of non-null type ' .
+                        '"' . Utils::printSafe($argType) . '" must not be null.',
                         $referenceNode
                     );
                 }
-            } elseif ($argumentValueNode instanceof VariableNode) {
-                $variableName = $argumentValueNode->name->value;
 
-                if ($variableValues !== null && array_key_exists($variableName, $variableValues)) {
-                    // Note: this does not check that this variable value is correct.
-                    // This assumes that this query has been validated and the variable
-                    // usage here is of the correct type.
-                    $coercedValues[$name] = $variableValues[$variableName];
-                } elseif ($argumentDefinition->defaultValueExists()) {
-                    $coercedValues[$name] = $argumentDefinition->defaultValue;
-                } elseif ($argType instanceof NonNull) {
+                if ($argumentValueNode instanceof VariableNode) {
+                    $variableName = $argumentValueNode->name->value;
                     throw new Error(
                         'Argument "' . $name . '" of required type "' . Utils::printSafe($argType) . '" was ' .
                         'provided the variable "$' . $variableName . '" which was not provided ' .
@@ -236,19 +257,38 @@ class Values
                         [$argumentValueNode]
                     );
                 }
-            } else {
-                $valueNode    = $argumentValueNode;
-                $coercedValue = AST::valueFromAST($valueNode, $argType, $variableValues);
-                if (Utils::isInvalid($coercedValue)) {
-                    // Note: ValuesOfCorrectType validation should catch this before
-                    // execution. This is a runtime check to ensure execution does not
-                    // continue with an invalid argument value.
-                    throw new Error(
-                        'Argument "' . $name . '" has invalid value ' . Printer::doPrint($valueNode) . '.',
-                        [$argumentValueNode]
-                    );
+
+                throw new Error(
+                    'Argument "' . $name . '" of required type ' .
+                    '"' . Utils::printSafe($argType) . '" was not provided.',
+                    $referenceNode
+                );
+            } elseif ($hasValue) {
+                if ($argumentValueNode instanceof NullValueNode) {
+                  // If the explicit value `null` was provided, an entry in the coerced
+                  // values must exist as the value `null`.
+                    $coercedValues[$name] = null;
+                } elseif ($argumentValueNode instanceof VariableNode) {
+                    $variableName = $argumentValueNode->name->value;
+                    Utils::invariant($variableValues !== null, 'Must exist for hasValue to be true.');
+                  // Note: This does no further checking that this variable is correct.
+                  // This assumes that this query has been validated and the variable
+                  // usage here is of the correct type.
+                    $coercedValues[$name] = $variableValues[$variableName] ?? null;
+                } else {
+                    $valueNode    = $argumentValueNode;
+                    $coercedValue = AST::valueFromAST($valueNode, $argType, $variableValues);
+                    if (Utils::isInvalid($coercedValue)) {
+                      // Note: ValuesOfCorrectType validation should catch this before
+                      // execution. This is a runtime check to ensure execution does not
+                      // continue with an invalid argument value.
+                        throw new Error(
+                            'Argument "' . $name . '" has invalid value ' . Printer::doPrint($valueNode) . '.',
+                            [$argumentValueNode]
+                        );
+                    }
+                    $coercedValues[$name] = $coercedValue;
                 }
-                $coercedValues[$name] = $coercedValue;
             }
         }